+++ src/app/pony/pony.component.ts
@@ -11,10 +11,12 @@ import { PonyModel } from '../models/pony.model';
 })
 export class PonyComponent {
   @Input({ required: true }) ponyModel!: PonyModel;
+  @Input() isRunning = false;
+
   @Output() readonly ponyClicked = new EventEmitter<PonyModel>();
 
   getPonyImageUrl(): string {
-    return `assets/images/pony-${this.ponyModel.color.toLowerCase()}.gif`;
+    return `assets/images/pony-${this.ponyModel.color.toLowerCase()}${this.isRunning ? '-running' : ''}.gif`;
   }
 
   clicked() {